requirements.txt
about-time==4.2.1
alive-progress==3.1.5
colour==0.1.5
grapheme==0.6.0
pillow==10.4.0
pyserial==3.5